requirements.txt
requests
cryptography
python-dotenv
-e git+git://github.com/djc/abna.git#egg=abna
